debt - don't let strings-util depend on map-util
