+++ src/vs/base/node/extfs.ts
@@ -5,15 +5,16 @@
 
 'use strict';
 
-import * as uuid from 'vs/base/common/uuid';
-import * as strings from 'vs/base/common/strings';
-import * as platform from 'vs/base/common/platform';
-import * as flow from 'vs/base/node/flow';
 import * as fs from 'fs';
 import * as paths from 'path';
-import { TPromise } from 'vs/base/common/winjs.base';
 import { nfcall } from 'vs/base/common/async';
+import { normalizeNFC } from 'vs/base/common/normalization';
+import * as platform from 'vs/base/common/platform';
+import * as strings from 'vs/base/common/strings';
+import * as uuid from 'vs/base/common/uuid';
+import { TPromise } from 'vs/base/common/winjs.base';
 import { encode, encodeStream } from 'vs/base/node/encoding';
+import * as flow from 'vs/base/node/flow';
 
 const loop = flow.loop;
 
@@ -21,7 +22,7 @@ export function readdirSync(path: string): string[] {
 	// Mac: uses NFD unicode form on disk, but we want NFC
 	// See also https://github.com/nodejs/node/issues/2165
 	if (platform.isMacintosh) {
-		return fs.readdirSync(path).map(c => strings.normalizeNFC(c));
+		return fs.readdirSync(path).map(c => normalizeNFC(c));
 	}
 
 	return fs.readdirSync(path);
@@ -36,7 +37,7 @@ export function readdir(path: string, callback: (error: Error, files: string[]) 
 				return callback(error, null);
 			}
 
-			return callback(null, children.map(c => strings.normalizeNFC(c)));
+			return callback(null, children.map(c => normalizeNFC(c)));
 		});
 	}
 
@@ -630,7 +631,7 @@ export function watch(path: string, onChange: (type: string, path: string) => vo
 				if (platform.isMacintosh) {
 					// Mac: uses NFD unicode form on disk, but we want NFC
 					// See also https://github.com/nodejs/node/issues/2165
-					file = strings.normalizeNFC(file);
+					file = normalizeNFC(file);
 				}
 			}
